requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Authlib==1.3.0
python-dotenv==1.0.1
stripe==7.12.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
email-validator==2.1.0.post1
python-jose==3.3.0
pytest==7.4.4
requests==2.31.0
PyYAML==6.0.1
docker==7.1.0
